change in q6rankings from float to positive small int

+++ rso_backend/competitions/models.py
@@ -494,7 +494,7 @@ class Q5EducatedParticipant(models.Model):
 
 
 class Q6TandemRanking(QBaseTandemRanking):
-    place = models.FloatField(
+    place = models.PositiveSmallIntegerField(
         verbose_name='Итоговое место по показателю 6'
     )
 
@@ -504,7 +504,7 @@ class Q6TandemRanking(QBaseTandemRanking):
 
 
 class Q6Ranking(QBaseRanking):
-    place = models.FloatField(
+    place = models.PositiveSmallIntegerField(
         verbose_name='Итоговое место по показателю 6'
     )
